requirements.txt
numpy
scipy
pytest
pytest-cov
tensorly
tensorly-torch
opt-einsum